Create App_Analyst.py
My point in this project is to find a mobile app that is profitable for App Store Play market. I am working as a data analyst
for a company that builds Android and IOS mobile apps, my job is to enable our team of enabling to make a data-driven decision with respect to the kind of apps they build. My goal in this project is to analyze data that will help the company's developers to understand what kind of apps are likely to attract more user.